+++ n3/bbox.h
@@ -4,20 +4,20 @@
 class BBox
 {
 	private:
-		GLfloat minX;
-		GLfloat maxX;
-		GLfloat minY;
-		GLfloat maxY;
+		int minX;
+		int maxX;
+		int minY;
+		int maxY;
 
 	public:
-		GLfloat getMinX(void);
-		GLfloat getMaxX(void);
-		GLfloat getMinY(void);
-		GLfloat getMaxY(void);
-		void setMinX(GLfloat);
-		void setMaxX(GLfloat);
-		void setMinY(GLfloat);
-		void setMaxY(GLfloat);
+		int getMinX(void);
+		int getMaxX(void);
+		int getMinY(void);
+		int getMaxY(void);
+		void setMinX(int);
+		void setMaxX(int);
+		void setMinY(int);
+		void setMaxY(int);
 		int pontoEstaDentro(VART::Point4D *);
 		VART::Point4D* getCentroObj();
 };
